refactor: move event classes to kafka/events/ from events/
