+++ common/src/main/java/com/turkcell/crm/common/CommonApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class CommonApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(CommonApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(CommonApplication.class, args);
+    }
 
 }
