+++ customerService/src/main/java/com/turkcell/crm/customer_service/business/concretes/IndividualCustomerManager.java
@@ -1,17 +1,17 @@
 package com.turkcell.crm.customer_service.business.concretes;
 
-import com.turkcell.crm.common.events.CustomerCreatedEvent;
+import com.turkcell.crm.common.kafka.events.CustomerCreatedEvent;
 import com.turkcell.crm.customer_service.business.abstracts.CustomerService;
 import com.turkcell.crm.customer_service.business.abstracts.IndividualCustomerService;
 import com.turkcell.crm.customer_service.business.dtos.requests.individual_customers.CreateIndividualCustomerRequest;
 import com.turkcell.crm.customer_service.business.dtos.requests.individual_customers.UpdateIndividualCustomerRequest;
 import com.turkcell.crm.customer_service.business.dtos.responses.individual_customers.*;
+import com.turkcell.crm.customer_service.business.kafka.producers.CustomerProducer;
 import com.turkcell.crm.customer_service.business.mappers.IndividualCustomerMapper;
 import com.turkcell.crm.customer_service.business.rules.IndividualCustomerBusinessRules;
 import com.turkcell.crm.customer_service.data_access.abstracts.IndividualCustomerRepository;
 import com.turkcell.crm.customer_service.entities.concretes.Customer;
 import com.turkcell.crm.customer_service.entities.concretes.IndividualCustomer;
-import com.turkcell.crm.customer_service.kafka.producers.CustomerProducer;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
