+++ customerService/src/main/java/com/turkcell/crm/customer_service/business/mappers/IndividualCustomerMapper.java
@@ -1,6 +1,6 @@
 package com.turkcell.crm.customer_service.business.mappers;
 
-import com.turkcell.crm.common.events.CustomerCreatedEvent;
+import com.turkcell.crm.common.kafka.events.CustomerCreatedEvent;
 import com.turkcell.crm.customer_service.business.dtos.requests.individual_customers.CreateIndividualCustomerRequest;
 import com.turkcell.crm.customer_service.business.dtos.requests.individual_customers.UpdateIndividualCustomerRequest;
 import com.turkcell.crm.customer_service.business.dtos.responses.individual_customers.*;
